test: split PDF hyperlink annotation cases
